+++ src/test/suite/utils/getMillisecondsRemainingInTrial.test.ts
@@ -1,0 +1,50 @@
+import * as assert from 'assert';
+import { TRIAL_LENGTH_IN_MILLISECONDS } from '../../../constants';
+import getMillisecondsRemainingInTrial from '../../../utils/getMillisecondsRemainingInTrial';
+
+const hourInMilliseconds = 3600000;
+suite('getMillisecondsRemainingInTrial', () => {
+  test('entire trial remaining', () => {
+    const time = new Date().getTime();
+    assert.strictEqual(
+      getMillisecondsRemainingInTrial(time, time, TRIAL_LENGTH_IN_MILLISECONDS),
+      172800000
+    );
+  });
+
+  test('1 day remaining', () => {
+    const time = new Date().getTime();
+    assert.strictEqual(
+      getMillisecondsRemainingInTrial(
+        time,
+        time - 24 * hourInMilliseconds,
+        TRIAL_LENGTH_IN_MILLISECONDS
+      ),
+      86400000
+    );
+  });
+
+  test('1 hour remaining', () => {
+    const time = new Date().getTime();
+    assert.strictEqual(
+      getMillisecondsRemainingInTrial(
+        time,
+        time - 47 * hourInMilliseconds,
+        TRIAL_LENGTH_IN_MILLISECONDS
+      ),
+      3600000
+    );
+  });
+
+  test('no time remaining', () => {
+    const time = new Date().getTime();
+    assert.strictEqual(
+      getMillisecondsRemainingInTrial(
+        time,
+        time - 48 * hourInMilliseconds,
+        TRIAL_LENGTH_IN_MILLISECONDS
+      ),
+      0
+    );
+  });
+});