+++ src/utils/getMillisecondsRemainingInTrial.ts
@@ -1,0 +1,11 @@
+export default function getMillisecondsRemainingInTrial(
+  currentTimeStamp: number,
+  trialStartedTimeStamp: number | unknown,
+  trialLengthInMilliseconds: number
+): number {
+  const previewTimeElapsed = currentTimeStamp - Number(trialStartedTimeStamp);
+
+  return trialStartedTimeStamp
+    ? trialLengthInMilliseconds - previewTimeElapsed
+    : 0;
+}